pii_protection/requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
apache-beam==2.39.0
astroid==2.12.5
async-timeout==4.0.2
attrs==22.1.0
cachetools==4.2.4
certifi==2023.5.7
charset-normalizer==2.0.12
cloudpickle==2.1.0
crcmod==1.7
DateTime==4.8
db-dtypes==1.0.4
decorator==5.1.1
dill==0.3.1.1
docopt==0.6.2
dos2unix==1
et==0.0.2
fastavro==1.5.1
fasteners==0.17.3
flatten-json==0.1.13
frozenlist==1.3.1
fsspec==2022.10.0
gcsfs==2022.10.0
google-api-core==1.31.6
google-api-python-client==2.51.0
google-apitools==0.5.31
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
google-cloud-bigquery==2.34.4
google-cloud-bigquery-storage==2.13.2
google-cloud-bigtable==1.7.2
google-cloud-core==2.3.1
google-cloud-datastore==1.15.5
google-cloud-dlp==3.7.1
google-cloud-language==1.3.2
google-cloud-pubsub==2.13.0
google-cloud-pubsublite==1.4.2
google-cloud-recommendations-ai==0.2.0
google-cloud-spanner==1.19.3
google-cloud-storage==2.4.0
google-cloud-videointelligence==1.16.3
google-cloud-vision==1.0.2
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.3
grpc-google-iam-v1==0.12.4
grpcio==1.46.3
grpcio-gcp==0.2.2
grpcio-status==1.46.3
hdfs==2.7.0
httplib2==0.19.1
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
multidict==6.0.2
ndjson==0.3.1
numpy==1.22.4
oauth2client==4.1.3
oauthlib==3.2.2
orjson==3.7.2
overrides==6.1.0
packaging==21.3
pandas==1.4.2
pandas-gbq==0.17.9
platformdirs==2.5.2
proto-plus==1.20.6
protobuf
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydata-google-auth==1.4.0
pydot==1.4.2
pylint==2.15.0
pymongo==3.12.3
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
tomli==2.0.1
tomlkit==0.11.4
typing-utils==0.1.0
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
wrapt==1.14.1
xlrd==2.0.1
yarl==1.8.1
zope.interface==5.5.2